+++ changes.md
@@ -4,6 +4,7 @@
 -   Added: `NativeEmailHandler` now available for logging
 -   Added: Details view for log entries
 -   Added: Priority for jobs, can be run on a high or low priority queue handler
+-   Added: Cron expressions to cron handlers
 -   Changed: Switched the logs table to MyISAM engine
 -   Changed: Jobs are running on the laravel queue